issue #2: Moodle 4.5 support
