+++ version.php
@@ -26,8 +26,8 @@
 defined('MOODLE_INTERNAL') || die();
 
 $plugin->component = 'filter_forumembed';
-$plugin->version   = 2023012500;
-$plugin->release   = '2023012500';
+$plugin->version   = 2025100300;
+$plugin->release   = '2025100300';
 $plugin->requires  = 2020061513;
 $plugin->maturity  = MATURITY_STABLE;
-$plugin->supported = [39, 401];
+$plugin->supported = [39, 405];
